+++ example/lib/extension.dart
@@ -1,17 +1,11 @@
 import 'package:flutter/material.dart';
 import 'package:intl/intl.dart';
 
-import 'enumerations.dart';
 import 'l10n/app_localizations.dart';
 import 'theme/app_colors.dart';
 import 'theme/app_theme_extension.dart';
 
 enum TimeStampFormat { parse_12, parse_24 }
-
-extension NavigationExtension on State {
-  void pushRoute(Widget page) =>
-      Navigator.of(context).push(MaterialPageRoute(builder: (context) => page));
-}
 
 extension NavigatorExtention on BuildContext {
   Future<T?> pushRoute<T>(Widget page) =>
@@ -25,52 +19,6 @@ extension NavigatorExtention on BuildContext {
 }
 
 extension DateUtils on DateTime {
-  String get weekdayToFullString {
-    switch (weekday) {
-      case DateTime.monday:
-        return "Monday";
-      case DateTime.tuesday:
-        return "Tuesday";
-      case DateTime.wednesday:
-        return "Wednesday";
-      case DateTime.thursday:
-        return "Thursday";
-      case DateTime.friday:
-        return "Friday";
-      case DateTime.saturday:
-        return "Saturday";
-      case DateTime.sunday:
-        return "Sunday";
-      default:
-        return "Error";
-    }
-  }
-
-  String get weekdayToAbbreviatedString {
-    switch (weekday) {
-      case DateTime.monday:
-        return "M";
-      case DateTime.tuesday:
-        return "T";
-      case DateTime.wednesday:
-        return "W";
-      case DateTime.thursday:
-        return "T";
-      case DateTime.friday:
-        return "F";
-      case DateTime.saturday:
-        return "S";
-      case DateTime.sunday:
-        return "S";
-      default:
-        return "Err";
-    }
-  }
-
-  int get totalMinutes => hour * 60 + minute;
-
-  TimeOfDay get timeOfDay => TimeOfDay(hour: hour, minute: minute);
-
   DateTime copyWith({
     int? year,
     int? month,
@@ -95,20 +43,9 @@ extension DateUtils on DateTime {
     return DateFormat(format).format(this);
   }
 
-  DateTime stringToDateWithFormat({
-    required String format,
-    required String dateString,
-  }) => DateFormat(format).parse(dateString);
-
   String getTimeInFormat(TimeStampFormat format) => DateFormat(
     'h:mm${format == TimeStampFormat.parse_12 ? " a" : ""}',
   ).format(this).toUpperCase();
-
-  bool compareWithoutTime(DateTime date) =>
-      day == date.day && month == date.month && year == date.year;
-
-  bool compareTime(DateTime date) =>
-      hour == date.hour && minute == date.minute && second == date.second;
 }
 
 extension ColorExtension on Color {
@@ -120,12 +57,20 @@ extension ColorExtension on Color {
   }
 }
 
-extension StringExt on String {
-  String get capitalized => toBeginningOfSentenceCase(this) ?? "";
-}
-
-extension ViewNameExt on CalendarView {
-  String get name => toString().split(".").last;
+extension TimeOfDayExtension on TimeOfDay {
+  /// Formats the time as a string (e.g., "2:30 PM" or "14:30")
+  String getTimeInFormat(TimeStampFormat format) {
+    if (format == TimeStampFormat.parse_12) {
+      final period = hour >= 12 ? 'PM' : 'AM';
+      final displayHour = hour == 0 ? 12 : (hour > 12 ? hour - 12 : hour);
+      final minuteStr = minute.toString().padLeft(2, '0');
+      return '$displayHour:$minuteStr $period';
+    } else {
+      final hourStr = hour.toString().padLeft(2, '0');
+      final minuteStr = minute.toString().padLeft(2, '0');
+      return '$hourStr:$minuteStr';
+    }
+  }
 }
 
 extension BuildContextExtension on BuildContext {